git subrepo pull subrepos/image-creation
subrepo:
  subdir:   "subrepos/image-creation"
  merged:   "8356e10"
upstream:
  origin:   "git@gitlab.internal.sanger.ac.uk:hgi/image-creation.git"
  branch:   "master"
  commit:   "8356e10"
git-subrepo:
  version:  "0.3.1"
  origin:   "https://github.com/Homebrew/brew.git"
  commit:   "664d0c6"

+++ subrepos/image-creation/kitchen_wrapper.sh
@@ -31,23 +31,26 @@ kitchen setup $KITCHEN_FLAGS> >(tee -a /tmp/stdout-${BUILD}.log) 2> >(tee -a /tm
 
 kitchen verify $KITCHEN_FLAGS> >(tee -a /tmp/stdout-${BUILD}.log) 2> >(tee -a /tmp/stderr-${BUILD}.log >&2) || {
   ERR=$?
-  echo "cat << EOF > /tmp/tmp_key.$$"
-  cat $KEYPAIR_FILENAME
-  echo "EOF"
-  echo "chmod 0700  /tmp/tmp_key.$$"
-  IP=`grep "Attaching floating IP" /tmp/stdout-${BUILD}.log | tail -1 | sed -e 's/^.*<//' -e 's/>$//' -e 's/ *//g'`
-  echo "ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null -o IdentityFile=/tmp/tmp_key.$$ -o IdentitiesOnly=yes ${IMAGE_USERNAME}@${IP} -i /tmp/tmp_key.$$"
-  echo "rm /tmp/tmp_key.$$"
-  let TOTAL_TIME=${SLEEPTIME:=1200}
-  let EACH_LOOP=TOTAL_TIME/100
-  for ((n=1;n<101;n++))
-  do
-    echo "$n of 100...."
-    sleep $EACH_LOOP
-  done
-  rm -f /tmp/stdout-${BUILD}.log  /tmp/stderr-${BUILD}.log
-  kitchen destroy
-  openstack keypair delete $KEYPAIR
+  if [ "$DEBUG" == "1" ]
+  then
+    echo "cat << EOF > /tmp/tmp_key.$$"
+    cat $KEYPAIR_FILENAME
+    echo "EOF"
+    echo "chmod 0700  /tmp/tmp_key.$$"
+    IP=`grep "Attaching floating IP" /tmp/stdout-${BUILD}.log | tail -1 | sed -e 's/^.*<//' -e 's/>$//' -e 's/ *//g'`
+    echo "ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null -o IdentityFile=/tmp/tmp_key.$$ -o IdentitiesOnly=yes ${IMAGE_USERNAME}@${IP} -i /tmp/tmp_key.$$"
+    echo "rm /tmp/tmp_key.$$"
+    let TOTAL_TIME=${SLEEPTIME:=1200}
+    let EACH_LOOP=TOTAL_TIME/100
+    for ((n=1;n<101;n++))
+    do
+      echo "$n of 100...."
+      sleep $EACH_LOOP
+    done
+    rm -f /tmp/stdout-${BUILD}.log  /tmp/stderr-${BUILD}.log
+    kitchen destroy
+    openstack keypair delete $KEYPAIR
+  fi
   exit $ERR
  }
 IP=`grep "Attaching floating IP" /tmp/stdout-${BUILD}.log | tail -1 | sed -e 's/^.*<//' -e 's/>$//' -e 's/ *//'`
